requirements.txt
fastapi[standard-no-fastapi-cloud-cli]
python-dotenv
openai
aiofiles
pydub
ffmpeg-python